_masters_/ml/requirements.txt
opencv-python==4.1.1.26
webcolors==1.3
scipy==0.15.1
matplotlib==3.1.1
keras==2.3.1
mtcnn==0.0.9
Pillow==6.2.0
tensorflow==1.4
